Removed the version number and added comments. Changed Package.
Drupal.org will add the version number when we create a tag release. Added section comments as seen in a lot of modules but not a necessity. Changed the package name to have fulfilment instead of shipping, because shipping section are meant for things like USPS, FEDEX etc. Also, went with the (name) schema that commerce has done with their core commerce modules.