Moved admin page code. Added 'file' key to menu item. Fixed a couple of small bugs with the admin form.
Admin settings form is now in its own file. Added a logic check on shipping methods to make sure that we have a shipping method available since we can't outright require a certain one. Added a message to the logic check if you don't have any shipping methods. Added the t() function around a description of a form element. Fixed the default value to be an empty array instead of a string for the checkboxes.